bugfix 404 ai feedback
